Create cpp file for argparser API
