+++ include/util/arg_parser.h
@@ -1,6 +1,6 @@
 /**
  * \file
- * \brief  Single-header for command-line argument parsing
+ * \brief  Class for command-line argument parsing
  */
 
 #ifndef INCLUDE_UTIL_ARG_PARSER_H_
@@ -10,12 +10,10 @@
 #include <array>
 #include <functional>
 #include <iomanip>
-#include <iostream>
 #include <memory>
 #include <optional>
 #include <set>
 #include <string>
-#include <string_view>
 #include <unordered_map>
 #include <utility>
 #include <variant>
@@ -28,6 +26,8 @@ class ArgumentParser;
 
 //! Unique instance to argument parser object
 using Parser = std::unique_ptr<ArgumentParser>;
+
+/* ********************************************************************************************** */
 
 //! Single argument option
 struct Argument {
@@ -45,18 +45,20 @@ struct Argument {
 //! List of mapped arguments to handle
 using ExpectedArguments = std::vector<Argument>;
 
+/* ********************************************************************************************** */
+
+//! Available type values an argument can have it
+struct Value : std::variant<std::monostate, bool, std::string> {
+  using std::variant<std::monostate, bool, std::string>::variant;
+
+  const bool& get_bool() const { return std::get<bool>(*this); }
+  const std::string& get_string() const { return std::get<std::string>(*this); }
+};
+
 /**
  * @brief Contains all arguments parsed from command-line
  */
 struct ParsedArguments {
-  //! Available type values to store
-  struct Value : std::variant<std::monostate, bool, std::string> {
-    using std::variant<std::monostate, bool, std::string>::variant;
-
-    const bool& get_bool() const { return std::get<bool>(*this); }
-    const std::string& get_string() const { return std::get<std::string>(*this); }
-  };
-
   //! Argument may contain or not a value associated
   using Arguments = std::unordered_map<std::string, std::optional<Value>>;
 
@@ -111,26 +113,7 @@ class ArgumentParser {
    * @brief Create a new ArgumentParser object
    * @param args List of expected arguments
    */
-  void Add(const ExpectedArguments& args) {
-    // Lambda to compare with choices from help argument
-    auto match_help = [](const std::string_view& choice) {
-      return choice == "-h" || choice == "--help";
-    };
-
-    // Insert expected arguments into internal cache
-    for (const auto& arg : args) {
-      // Check if argument matches some fields from default help argument
-      if (auto matched = std::find_if(arg.choices.begin(), arg.choices.end(), match_help);
-          arg.name == "help" || matched != std::end(arg.choices)) {
-        throw parsing_error("Cannot override default help text");
-      }
-
-      // TODO: filter arg.choices to match a single char OR word
-
-      if (auto [dummy, inserted] = expected_arguments_.insert(arg); !inserted)
-        throw parsing_error("Cannot configure duplicated argument");
-    }
-  }
+  void Add(const ExpectedArguments& args);
 
  public:
   /**
@@ -138,16 +121,7 @@ class ArgumentParser {
    * @param args List of expected arguments
    * @return Parser unique instance
    */
-  static Parser Configure(const ExpectedArguments& args) {
-    // Simply extend the ArgumentParser class, as we do not want to expose the default constructor,
-    // neither do we want to use std::make_unique explicitly calling operator new()
-    struct MakeUniqueEnabler : public ArgumentParser {};
-    auto parser = std::make_unique<MakeUniqueEnabler>();
-
-    if (!args.empty()) parser->Add(args);
-
-    return parser;
-  }
+  static Parser Configure(const ExpectedArguments& args);
 
   /**
    * @brief Destroy an ArgumentParser object
@@ -170,55 +144,7 @@ class ArgumentParser {
    * @return A map containing all parsed arguments where key is the argument identifier and value is
    * the value read for that argument
    */
-  ParsedArguments Parse(int count, char** values) const {
-    ParsedArguments opts;
-    if (count == 1) return opts;
-
-    int index = 1;
-    while (index < count) {
-      std::string argument{values[index]};
-
-      if (argument == "-h" || argument == "--help") {
-        PrintHelp();
-        throw parsing_error("Received command to print helper");
-      }
-
-      // Lambda to match argument choice
-      auto match_choice = [&argument](const Argument& arg) {
-        return std::find(arg.choices.begin(), arg.choices.end(), argument) != arg.choices.end();
-      };
-
-      // Find match in choices from expected arguments
-      auto found =
-          std::find_if(expected_arguments_.begin(), expected_arguments_.end(), match_choice);
-
-      if (found == expected_arguments_.end()) {
-        PrintError(argument);
-        throw parsing_error("Received unexpected argument");
-      }
-
-      // Argument does not expect a value
-      if (found->is_empty) {
-        opts[found->name] = true;
-      } else {
-        // Parse value for argument
-        std::string value = values[++index];
-
-        // Get value for expected argument
-        if (value.rfind('-', 0) == 0 || value.empty()) {
-          PrintError(argument, value);
-          throw parsing_error("Received unexpected value for argument");
-        }
-
-        // Everything is fine, should include into opts
-        opts[found->name] = value;
-      }
-
-      ++index;
-    }
-
-    return opts;
-  }
+  ParsedArguments Parse(int count, char** values) const;
 
   /* ******************************************************************************************** */
   //! Utility
@@ -231,64 +157,25 @@ class ArgumentParser {
    * @param args Expected arguments
    * @return Length value
    */
-  int GetBiggestLength(const Filtered& args) const {
-    int length = 0;
-
-    for (const auto& arg : args) {
-      int choice_length = 0;
-      for (const auto& choice : arg.choices) choice_length += (int)choice.size();
-
-      if (length < choice_length) length = choice_length;
-    }
-
-    return length;
-  }
+  int GetBiggestLength(const Filtered& args) const;
 
   /**
-   * @brief Utility method to print a CLI helper based on expected arguments
+   * @brief Utility method to print a CLI helper based on expected arguments and abort parsing
    */
-  void PrintHelp() const {
-    std::cout << "spectrum\n\n";
-    std::cout << "A music player with a simple and intuitive terminal user interface.\n\n";
-    std::cout << "Options:";
-
-    auto biggest_choice = GetBiggestLength(expected_arguments_);
-
-    for (const auto& arg : expected_arguments_) {
-      std::cout << "\n\t";
-
-      // Append choices into a single string
-      std::string choices;
-      for (const auto& choice : arg.choices) choices.append(choice + ", ");
-
-      // Remove last comma+space
-      if (choices.size() > 1) choices.erase(choices.size() - 2);
-
-      std::cout << std::setw(biggest_choice) << std::left << choices;
-      std::cout << "\t" << arg.description;
-    }
-    std::cout << "\n";
-  }
+  void PrintHelpAndThrow() const;
 
   /**
-   * @brief Utility method to print error on CLI
+   * @brief Utility method to print error on CLI and abort parsing
    * @param parsed Argument parsed from command-line
    */
-  void PrintError(const std::string& parsed) const {
-    if (parsed.empty())
-      std::cout << "spectrum: empty option\n";
-    else
-      std::cout << "spectrum: invalid option [" << parsed << "]\n";
-  }
+  void PrintErrorAndThrow(const std::string& parsed) const;
 
   /**
-   * @brief Utility method to print error on CLI
+   * @brief Utility method to print error on CLI and abort parsing
    * @param argument Argument parsed from command-line
    * @param value Value parsed from command-line
    */
-  void PrintError(const std::string& argument, const std::string& value) const {
-    std::cout << "spectrum: invalid value(" << value << ") for option [" << argument << "]\n";
-  }
+  void PrintErrorAndThrow(const std::string& argument, const std::string& value) const;
 
   /* ******************************************************************************************** */
   //! Variables
@@ -296,7 +183,8 @@ class ArgumentParser {
   //!< Expected arguments for command-line parsing
   Filtered expected_arguments_ = {Argument{.name = "help",
                                            .choices = {"-h", "--help"},
-                                           .description = "Display this help text and exit"}};
+                                           .description = "Display this help text and exit",
+                                           .is_empty = false}};
 };
 
 }  // namespace util
